[Core] Ensure sane repositories
